Add optional timeout for commands
This prevents infinite blocking when BT chip is not present or is not
responding to the commands. Usually, functions check for pending data
using UART peek callback first, but in some cases data is just awaited
(sl_bt_wait_response, sli_bgapi_get_event), causing infinite blocking
when BT chip is not present or not responding.