Fix phase calculation during enhanced-ack generation
There was a bug when calculating the phase value while generating the enhanced ack. We were not considering the correct packet length value, which was causing the incorrect phase value. This later was causing the next Rx-window calculation to be wrong for WED after receiving the enhanced ack.

Also backup and restore the original received packet bytes length if it is exceeded the expected receive data bytes. This is needed for correct phase calculation in the later part of the enh. ack generation code.