Set ccaBackoff to some constant value to have a predictable radio warmup time for schedule tx
If the CCA backoff period is set to 0 for the fixed CCA, and the PROTIMER can't get the radio on in 0 time for the CCA check, so the MCU turns the radio on when the scheduled TX is submitted. To avoid this, set .ccaBackoff to some value and take that into account when scheduling the time for the TX with single-CCA to start. This will allow the PROTIMER to turn on the radio in time for the CCA check. Also with zero CCA backoff period, it has been observed that time between RxSearch to RxToTx ready is variable if ccaBackoff period is zero, so it could affect the actual schedule tx time and cause unexpected Tx drift.

Rail events with approximate timings while schedule transmit with a single CCA look like following
RxWarm <----~102usec----> RxSearch <----~150usec---> RxToTx <---~15usec---> LBTSuccess <----~102usec----> TxStart (0xFC)